+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2021 SamSDC
+Copyright (c) 2021 hr-rfp53-Starfire
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
