--- README copy.md
@@ -1,58 +1,0 @@
-# starfire-project-catwalk
-### Table of Contents
-1. [General Info](#🌴-General-Info)
-2. [Demo](#✨-Demo)
-3. [Technologies](#🧪-Technologies)
-4. [Installation](#🚀-Installation)
-5. [FAQs](#FAQS)
-6. [Contributors](#🤝-Contributors)
-
-
-### General Info
-Starfire-project-catwalk is a client-facing retail web-portal that allows customers to browse items in a retail catalog.
-
-### ✨ Demo
-#### Overview
-![](image-gallery.gif)
-![](add-to-cart.gif)
-
-#### Question & Answers
-![](search-highlight.gif)
-![](answer-form.gif)
-
-#### Rating & Reviews
-![](filters-demo.gif)
-![](review-form.gif)
-
-### 🧪 Technologies
-* Axios : Version 0.21.1,
-* Cloudinary : Version 1.26.0,
-* Cssloader : Version 1.1.1,
-* Express : Version 4.17.1,
-* Express-form-data : Version 2.0.17,
-* Identity-obj-proxy : Version 3.0.0,
-* Loader : Version 2.1.1,
-* Luxon : Version 1.27.0,
-* Nodemon: Version 2.0.7,
-* React : Version 17.0.2,
-* React-dom : Version 17.0.2,
-* React-lazyload : Version 3.2.0,
-* React-router-dom : Version 5.2.0
-
-### 🚀 Installation
-```
-$ git clone https://github.com/hr-rfp53-Starfire/starfire-project-catwalk.git
-$ cd ../starfire-project-catwalk
-$ npm install
-$ npm run build:dev
-$ npm start
-```
-
-### FAQs
-
-
-### 🤝 Contributors
-- [Tea Skela](https://github.com/tskela)
-- [Laura Kamphaus](https://github.com/lkamphaus)
-- [Jake Melnick](https://github.com/JacobMelnick)
-
