+++ client/src/components/reviews/Reviews.module.css
@@ -1,1 +1,1 @@
-.gridContainer{color:var(--text-color);font-size:18px;display:grid;width:75%;padding-bottom:50px;margin:auto;column-gap:5%;grid-template-columns:1fr 1.6fr;grid-template-rows:auto;grid-template-areas:"Product-Breakdown Reviews-List ." ". . ." ". . ."}.reviewsList{grid-area:Reviews-List;max-height:800px;min-height:55vh;overflow-y:auto;overflow-x:hidden}.productBreakdown{grid-area:Product-Breakdown;padding-top:35px;text-align:left}.reviewTile{width:95%;border-bottom:1px solid grey;padding-bottom:20px;padding-top:10px;margin-bottom:20px}.reviewButtons{display:in-line;margin-top:20px;margin-right:20px;margin-bottom:0;padding:20px;color:#fff;font-size:15px;line-height:20px;background-color:#26a7ba}.reviewSummary{font-size:20px;font-weight:700;margin-bottom:10px}.userDate{font-size:12px;text-align:right;opacity:80%;margin-bottom:15px;padding-right:5px}.thumbnails{display:inline}.smallImg{display:inline;margin-top:10px;margin-right:10px;}.showMore{text-decoration:underline;cursor:pointer}.modal{background-color:rgba(0,0,0,.5);position:fixed;height:100%;width:100%;top:0;left:0;display:flex;align-items:center;justify-content:center}.exit{color:#d3d3d3;margin:10px;text-align:right;cursor:pointer;display:block}.reviewForm{color:grey;margin:20px}.formTitle{margin-bottom:10px;margin-top:10px}.fields{font-family:Roboto,sans-serif}.dropDown{-webkit-appearance:none;color:grey;border:none;background-color:#f5efed;font-family:Roboto,sans-serif;font-size:18px;text-decoration:underline}.stars{height:20px;width:20px;display:inline-block;vertical-align:top;margin-top:15px}.star{height:20px;width:20px}.ratingFilters{width:100%}.bar{padding:0;display:inline-block;vertical-align:middle;background-color:#d3d3d3;margin:0;height:20px;width:80%;border-radius:50px;padding-bottom:5px}.filler{color:rgba(255,255,255,.75);text-align:center;background:#d96c06;height:100%;border-radius:inherit;padding-bottom:5px}.ratingBar{margin:20px 0;transition:transform .3s;cursor:pointer}.ratingBar:hover{transform:scale(1.1)}.response{padding:20px;margin:10px 0;background-color:#d3d3d3}.characteristicRow{width:90%;display:flex}.characteristic{margin-left:auto;margin-right:0}.characteristicSelection{padding:5px;font-size:14px;color:#d96c06}.icon{height:100%;border-radius:inherit;color:var(--icon-color)}#reviews{scroll-behavior:smooth}
+.gridContainer{color:var(--text-color);font-size:18px;display:grid;width:75%;padding-bottom:50px;margin:auto;column-gap:5%;grid-template-columns:1fr 1.6fr;grid-template-rows:auto;grid-template-areas:"Product-Breakdown Reviews-List ." ". . ." ". . ."}.reviewsList{grid-area:Reviews-List;max-height:800px;min-height:55vh;overflow-y:auto;overflow-x:hidden}.productBreakdown{grid-area:Product-Breakdown;padding-top:35px;text-align:left}.reviewTile{width:95%;border-bottom:1px solid grey;padding-bottom:20px;padding-top:10px;margin-bottom:20px}.reviewButtons{display:in-line;margin-top:20px;margin-right:20px;margin-bottom:0;padding:20px;color:#fff;font-size:15px;line-height:20px;background-color:#26a7ba}.reviewSummary{font-size:20px;font-weight:700;margin-bottom:10px}.userDate{font-size:12px;text-align:right;opacity:80%;margin-bottom:15px;padding-right:5px}.thumbnails{display:inline}.smallImg{display:inline;margin-top:10px;margin-right:10px;height:70px;width:70px}.showMore{text-decoration:underline;cursor:pointer}.modal{background-color:rgba(0,0,0,.5);position:fixed;height:100%;width:100%;top:0;left:0;display:flex;align-items:center;justify-content:center}.exit{color:#d3d3d3;margin:10px;text-align:right;cursor:pointer;display:block}.reviewForm{color:grey;margin:20px}.formTitle{margin-bottom:10px;margin-top:10px}.fields{font-family:Roboto,sans-serif}.dropDown{-webkit-appearance:none;color:grey;border:none;background-color:#f5efed;font-family:Roboto,sans-serif;font-size:18px;text-decoration:underline}.stars{height:20px;width:20px;display:inline-block;vertical-align:top;margin-top:15px}.star{height:20px;width:20px}.ratingFilters{width:100%}.bar{padding:0;display:inline-block;vertical-align:middle;background-color:#d3d3d3;margin:0;height:20px;width:80%;border-radius:50px;padding-bottom:5px}.filler{color:rgba(255,255,255,.75);text-align:center;background:#d96c06;height:100%;border-radius:inherit;padding-bottom:5px}.ratingBar{margin:20px 0;transition:transform .3s;cursor:pointer}.ratingBar:hover{transform:scale(1.1)}.response{padding:20px;margin:10px 0;background-color:#d3d3d3}.characteristicRow{width:90%;display:flex}.characteristic{margin-left:auto;margin-right:0}.characteristicSelection{padding:5px;font-size:14px;color:#d96c06}.icon{height:100%;border-radius:inherit;color:var(--icon-color)}#reviews{scroll-behavior:smooth}
