+++ config.js
@@ -1,3 +1,7 @@
 module.exports = {
-    TOKEN : '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'
-};
+    TOKEN: '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',
+    API_KEY: '698854889886283',
+    API_SECRET: 'O-emEIlAiASv5Z3_ZXbuRaXWt-U',
+    CLOUD_NAME: 'dkrtpalhl',
+    FILESTACK_API: 'AWs7LcKPWTNKPac9vsWCQz'
+  };
