--- webpack.common.js
@@ -1,55 +1,0 @@
-var path = require("path");
-var SRC_DIR = path.join(__dirname, "/client/src");
-var DIST_DIR = path.join(__dirname, "/client/dist");
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-
-
-module.exports = {
-  entry: `${SRC_DIR}/index.jsx`,
-  mode: "development",
-  output: {
-    filename: 'bundle.js',
-    path: DIST_DIR,
-    publicPath: '/'
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      title: 'Catwalk',
-      template: './template.html'
-    }),
-  ],
-  module: {
-    rules: [
-      {
-        test: /\.jsx?/,
-        loader: "babel-loader",
-        options: {
-          presets: ["@babel/preset-env", "@babel/preset-react"],
-          plugins: ["@babel/plugin-syntax-jsx"]
-        },
-      },
-      {
-        test: /\.css$/,
-        use: [
-          {
-            loader: 'style-loader'
-          },
-          {
-            loader: 'css-loader',
-            options: {
-              modules: true,
-            }
-          }
-        ]
-      },
-      {
-        test: /\.(jpg|png)$/,
-        use: [
-          {
-            loader: 'file-loader'
-          },
-        ]
-      }
-    ],
-  },
-};
