--- webpack.dev.js
@@ -1,11 +1,0 @@
-const { merge } = require('webpack-merge');
-const common = require('./webpack.common.js');
-
-module.exports = merge(common, {
-  mode: 'development',
-//   devtool: 'inline-source-map',
-  devServer: {
-    contentBase: './client/dist',
-  },
-  
-});
